Add forest_neighbor folder with files for element and face neighbor
